updatin models to new arch, adding base VisionModule
