+++ easydel/modules/gemma2/modeling_gemma2_flax.py
@@ -279,9 +279,9 @@ class FlaxGemma2Attention(FlaxAttentionModule):
 
 		if self.has_variable("cache", "cached_key") or init_cache:
 			key_states, value_states, attention_mask = self._concatenate_to_cache(
+				query_states,
 				key_states,
 				value_states,
-				query_states,
 				attention_mask,
 			)
