+++ easydel/modules/gpt_j/__init__.py
@@ -15,7 +15,7 @@
 from easydel.modules.gpt_j.gpt_j_configuration import GPTJConfig
 from easydel.modules.gpt_j.modeling_gpt_j_flax import (
 	FlaxGPTJForCausalLM,
-	FlaxGPTJForCausalLMModule,
 	FlaxGPTJModel,
-	FlaxGPTJModule,
 )
+
+__all__ = "FlaxGPTJForCausalLM", "FlaxGPTJModel", "GPTJConfig"
