+++ easydel/modules/internlm2/__init__.py
@@ -15,9 +15,13 @@
 from easydel.modules.internlm2.internlm2_configuration import InternLM2Config
 from easydel.modules.internlm2.modeling_internlm2_flax import (
 	FlaxInternLM2ForCausalLM,
-	FlaxInternLM2ForCausalLMModule,
 	FlaxInternLM2ForSequenceClassification,
-	FlaxInternLM2ForSequenceClassificationModule,
 	FlaxInternLM2Model,
-	FlaxInternLM2Module,
+)
+
+__all__ = (
+	"FlaxInternLM2ForCausalLM",
+	"FlaxInternLM2ForSequenceClassification",
+	"FlaxInternLM2Model",
+	"InternLM2Config",
 )
