+++ easydel/modules/internlm2/modeling_internlm2_flax.py
@@ -21,53 +21,33 @@ import jax
 import jax.numpy as jnp
 from einops import rearrange
 from flax import linen as nn
-from flax.core.frozen_dict import FrozenDict, freeze, unfreeze
-from flax.linen import Dense, combine_masks
+from flax.linen import Dense
 from flax.linen import partitioning as nn_partitioning
-from flax.traverse_util import flatten_dict, unflatten_dict
-from jax import lax
 from jax.sharding import PartitionSpec
 
 from easydel.layers.attention import FlaxAttentionModule, FlexibleAttentionModule
 from easydel.layers.norms import RMSNorm
+from easydel.layers.rotary_embedding import get_rope
 from easydel.modules.factory import register_module
 
 # easydel.modules
 from easydel.modules.flax_modeling_utils import (
 	ACT2FN,
-	apply_rotary_pos_emb,
 	block_wise_ffn,
 	control_mlp_sharding,
 	get_dot_general_by_bits,
 	get_gradient_checkpoint_policy,
-	precompute_frequencies,
 	with_sharding_constraint,
 )
 from easydel.modules.internlm2.internlm2_configuration import (
 	InternLM2Config as InternLM2Config,
 )
-from easydel.modules.internlm2.kernels import internlm2_mlp_pallas
 from easydel.modules.modeling_flax_outputs import (
 	FlaxBaseModelOutput,
 	FlaxCausalLMOutput,
 	FlaxSequenceClassifierOutput,
 )
-from easydel.modules.modeling_utils import EasyDeLBaseModule
-
-
-class FlaxInternLM2Embedding(nn.Module):
-	dtype: jnp.dtype = jnp.float32
-
-	def __call__(self, query, key, frequencies, position_ids):
-		sin, cos = frequencies
-
-		sin = sin[position_ids][:, None, :, :]
-		cos = cos[position_ids][:, None, :, :]
-
-		key = apply_rotary_pos_emb(key, sin, cos)
-		query = apply_rotary_pos_emb(query, sin, cos)
-
-		return query.astype(self.dtype), key.astype(self.dtype)
+from easydel.modules.modeling_utils import wrap_easydel_module
 
 
 class FlaxInternLM2Attention(FlaxAttentionModule):
@@ -118,7 +98,6 @@ class FlaxInternLM2Attention(FlaxAttentionModule):
 			**get_dot_general_by_bits(self.config.bits, self.config.easy_method),
 		)
 
-		self.rotary = FlaxInternLM2Embedding(self.dtype)
 		self.attention_performer = FlexibleAttentionModule(
 			num_q_heads=self.config.num_attention_heads,
 			num_kv_heads=self.config.num_key_value_heads,
@@ -133,45 +112,35 @@ class FlaxInternLM2Attention(FlaxAttentionModule):
 			base_config=self.config,
 		)
 
-	def _merge_heads(self, hidden_states):
-		"""
-		Merges the attention heads into a single hidden state tensor.
+		initial_rope_kwargs = dict(rope_type="default")
+		if config.rope_scaling is not None:
+			scaling_type = config.rope_scaling.get("rope_type", None)
+			scaling_type = config.rope_scaling.get("type", scaling_type)
+			scaling_factor = config.rope_scaling.get("factor")
+			low_freq_factor = config.rope_scaling.get("low_freq_factor", None)
+			high_freq_factor = config.rope_scaling.get("high_freq_factor", None)
+			original_max_position_embeddings = config.rope_scaling.get(
+				"original_max_position_embeddings", None
+			)
+			initial_rope_kwargs = dict(
+				scaling_factor=scaling_factor,
+				rope_type=scaling_type,
+				low_freq_factor=low_freq_factor,
+				high_freq_factor=high_freq_factor,
+				original_max_position_embeddings=original_max_position_embeddings,
+			)
 
-		Args:
-		    hidden_states (chex.Array): The hidden states with separate head dimensions.
-
-		Returns:
-		    chex.Array: The hidden states with merged head dimensions.
-		"""
-		return hidden_states.reshape(hidden_states.shape[:2] + (-1,))
-
-	def apply_rotary(self, query, key, frequencies, position_ids):
-		"""
-		Applies rotary positional embeddings to the query and key tensors.
-
-		Args:
-		    query (chex.Array): Query tensor.
-		    key (chex.Array): Key tensor.
-		    frequencies (Tuple[chex.Array, chex.Array]): Tuple containing cosine and sine components for rotary embeddings.
-		    position_ids (chex.Array): Position indices for the tokens.
-
-		Returns:
-		    Tuple[chex.Array, chex.Array]: The modified query and key tensors after applying rotary embeddings.
-		"""
-
-		query, key = self._transpose_sequence_head(query, key)
-		query, key = self.rotary(
-			position_ids=position_ids,
-			query=query,
-			key=key,
-			frequencies=frequencies,
+		self.rotary = get_rope(
+			head_size=self.head_dim,
+			rotary_dim=self.head_dim,
+			max_position=self.config.granted_freq_max_position_embedding,
+			base=config.rope_theta,
+			rope_scaling=initial_rope_kwargs,
 		)
-		return self._transpose_sequence_head(query, key)
 
 	def __call__(
 		self,
 		hidden_states: chex.Array,
-		frequencies: Tuple[chex.Array, chex.Array],
 		attention_mask: chex.Array,
 		position_ids: chex.Array,
 		causal_mask: chex.Array,
@@ -186,7 +155,6 @@ class FlaxInternLM2Attention(FlaxAttentionModule):
 
 		Args:
 		    hidden_states (chex.Array): Input hidden states.
-		    frequencies (Tuple[chex.Array, chex.Array]): Cosine and sine components for rotary embeddings.
 		    attention_mask (chex.Array): Mask to apply on the attention scores.
 		    position_ids (chex.Array): Position indices for the tokens.
 		    causal_mask (chex.Array): Causal mask for ensuring autoregressive behavior.
@@ -198,8 +166,6 @@ class FlaxInternLM2Attention(FlaxAttentionModule):
 		Returns:
 		    Tuple[chex.Array, chex.Array]: A tuple containing the attention output and the attention weights.
 		"""
-		batch_size = hidden_states.shape[0]
-
 		qkv_states = rearrange(
 			self.wqkv(hidden_states),
 			"b q (h gs d) -> b q h gs d",
@@ -211,50 +177,33 @@ class FlaxInternLM2Attention(FlaxAttentionModule):
 		query_states = rearrange(query_states, "b q h gs d -> b q (h gs) d")
 		key_states = qkv_states[..., -2, :]
 		value_states = qkv_states[..., -1, :]
-		query_states, key_states = self.apply_rotary(
+		query_states, key_states = self.rotary(
 			query=query_states,
 			key=key_states,
-			position_ids=position_ids,
-			frequencies=frequencies,
+			positions=position_ids,
 		)
 
 		query_length, key_length = query_states.shape[1], key_states.shape[1]
-
-		if self.has_variable("cache", "cached_key"):
-			mask_shift = self.variables["cache"]["cache_index"]
-			max_decoder_length = self.variables["cache"]["cached_key"].shape[1]
-			causal_mask = lax.dynamic_slice(
-				causal_mask,
-				(0, 0, mask_shift, 0),
-				(1, 1, query_length, max_decoder_length),
-			)
-		else:
-			causal_mask = causal_mask[:, :, :query_length, :key_length]
-
-		batch_size = hidden_states.shape[0]
-		causal_mask = jnp.broadcast_to(causal_mask, (batch_size,) + causal_mask.shape[1:])
-		if attention_mask.ndim == 2:
-			attention_mask = jnp.expand_dims(attention_mask, axis=(-3, -2))
-		attention_mask = jnp.broadcast_to(attention_mask, causal_mask.shape)
-		attention_mask = combine_masks(attention_mask, causal_mask, fcm_mask)
 
 		dropout_rng = None
 
 		if not deterministic and self.config.attention_dropout > 0.0:
 			dropout_rng = self.make_rng("dropout")
 
-		if self.has_variable("cache", "cached_key") or init_cache:
-			key_states, value_states, attention_mask = self._concatenate_to_cache(
-				key_states,
-				value_states,
-				query_states,
-				attention_mask,
-			)
-
-		attention_bias = lax.select(
-			attention_mask > 0,
-			jnp.full(attention_mask.shape, 0.0).astype(self.dtype),
-			jnp.full(attention_mask.shape, jnp.finfo(self.dtype).min).astype(self.dtype),
+		(
+			query_states,
+			key_states,
+			value_states,
+			attention_mask,
+			attention_bias,
+		) = self.concatenate_to_cache(
+			init_cache=init_cache,
+			query=query_states,
+			key=key_states,
+			value=value_states,
+			attention_mask=attention_mask,
+			causal_mask=causal_mask,
+			fcm_mask=fcm_mask,
 		)
 
 		query_length, key_length = query_states.shape[1], key_states.shape[1]
@@ -332,30 +281,7 @@ class FlaxInternLM2MLP(nn.Module):
 		    A tensor that is the result of applying a dropout function
 		    to x
 		"""
-
 		x = control_mlp_sharding(x, self.config.partition_axis)
-
-		if (
-			self.config.hardware_abstraction
-			and self.w2.variables.get("params", None) is not None
-		):
-			return jax.vmap(
-				functools.partial(
-					internlm2_mlp_pallas,
-					act_fn=self.act_fn,
-					blocksize_k=self.config.pallas_k_block_size,
-					blocksize_m=self.config.pallas_m_block_size,
-					blocksize_n=self.config.pallas_n_block_size,
-					prod_dtype=self.dtype,
-					precision=self.precision,
-				),
-				in_axes=(0, None, None, None),
-			)(
-				x,
-				self.w1.variables["params"]["kernel"],
-				self.w2.variables["params"]["kernel"],
-				self.w3.variables["params"]["kernel"],
-			)
 		return self.w2(self.act_fn(self.w1(x)) * self.w3(x))
 
 
@@ -371,7 +297,7 @@ class FlaxInternLM2Block(nn.Module):
 		if self.config.gradient_checkpointing != "":
 			attn_block = nn_partitioning.remat(
 				FlaxInternLM2Attention,
-				static_argnums=(1, 3, 4, 6, 7, 8),
+				static_argnums=(3, 4, 5, 6, 7),
 				policy=get_gradient_checkpoint_policy(self.config.gradient_checkpointing),
 			)
 			mlp_block = nn_partitioning.remat(
@@ -409,7 +335,6 @@ class FlaxInternLM2Block(nn.Module):
 	def __call__(
 		self,
 		hidden_states: chex.Array,
-		frequencies: Tuple[chex.Array, chex.Array],
 		attention_mask: chex.Array,
 		position_ids: chex.Array,
 		causal_mask: chex.Array,
@@ -424,7 +349,6 @@ class FlaxInternLM2Block(nn.Module):
 
 		Args:
 		    hidden_states (chex.Array): Input hidden states.
-		    frequencies (Tuple[chex.Array, chex.Array]): Cosine and sine components for rotary embeddings.
 		    attention_mask (chex.Array): Mask to apply on the attention scores.
 		    position_ids (chex.Array): Position indices for the tokens.
 		    causal_mask (chex.Array): Causal mask for ensuring autoregressive behavior.
@@ -438,7 +362,6 @@ class FlaxInternLM2Block(nn.Module):
 		"""
 		attn_outputs = self.attention(
 			self.attention_norm(hidden_states),
-			frequencies,
 			attention_mask,
 			position_ids,
 			causal_mask,
@@ -469,246 +392,6 @@ class FlaxInternLM2Block(nn.Module):
 		hidden_states = hidden_states + feed_forward_hidden_states
 
 		return (hidden_states,) + attn_outputs[1:]
-
-
-class FlaxInternLM2PreTrainedModel(EasyDeLBaseModule):
-	"""
-	Base class for InternLM2 models providing initialization and configuration.
-
-	Attributes:
-	    config_class (InternLM2Config): The configuration class for the model.
-	    module_class (nn.Module): The class representing the model's architecture.
-	    base_model_prefix (str): The prefix for the base model parameters.
-	"""
-
-	config_class = InternLM2Config
-	base_model_prefix = "model"
-	module_class: nn.Module = None
-
-	def __init__(
-		self,
-		config: InternLM2Config,
-		dtype: jnp.dtype = jnp.bfloat16,
-		param_dtype: jnp.dtype = jnp.bfloat16,
-		precision: Optional[jax.lax.Precision] = None,
-		input_shape: Tuple[int, int] = (1, 1),
-		seed: int = 0,
-		_do_init: bool = False,
-		**kwargs,
-	):
-		"""
-		Initializes the pre-trained model with the given configuration.
-
-		Args:
-		    config (InternLM2Config): Configuration for the model.
-		    dtype (jnp.dtype): Data type for computations.
-		    param_dtype (jnp.dtype): Data type for model parameters.
-		    precision (Optional[jax.lax.Precision]): Precision setting for JAX operations.
-		    input_shape (Tuple[int, int]): Shape of the input tensor.
-		    seed (int): Seed for random number generation.
-		    _do_init (bool): If True, initialize model weights.
-		    **kwargs: Additional keyword arguments.
-		"""
-		module = self.module_class(
-			config=config,
-			dtype=dtype,
-			param_dtype=param_dtype,
-			precision=precision,
-			**kwargs,
-		)
-		super().__init__(
-			dtype=dtype,
-			_do_init=_do_init,
-			module=module,
-			config=config,
-			input_shape=input_shape,
-			seed=seed,
-		)
-
-	def init_weights(
-		self,
-		rng: jax.random.PRNGKey,
-		input_shape: Tuple,
-		params: FrozenDict = None,
-	) -> FrozenDict:
-		"""
-		Initializes the model weights.
-
-		Args:
-		    rng (jax.random.PRNGKey): Random number generator key.
-		    input_shape (Tuple): Shape of the input tensor for initializing weights.
-		    params (FrozenDict, optional): Existing parameters to initialize with.
-
-		Returns:
-		    FrozenDict: Initialized model parameters.
-		"""
-		input_ids = jnp.zeros(input_shape, dtype="i4")
-		attention_mask = jnp.ones_like(input_ids)
-		position_ids = jnp.broadcast_to(
-			jnp.arange(jnp.atleast_2d(input_ids).shape[-1]), input_shape
-		)
-		params_rng, dropout_rng = jax.random.split(rng)
-		rngs = {"params": params_rng, "dropout": dropout_rng}
-
-		if self.config.add_cross_attention:
-			encoder_hidden_states = jnp.zeros(input_shape + (self.config.hidden_size,))
-			encoder_attention_mask = attention_mask
-			module_init_outputs = self.module.init(
-				rngs,
-				input_ids,
-				attention_mask,
-				position_ids,
-				encoder_hidden_states,
-				encoder_attention_mask,
-				return_dict=False,
-			)
-		else:
-			module_init_outputs = self.module.init(
-				rngs,
-				input_ids=input_ids,
-				attention_mask=attention_mask,
-				position_ids=position_ids,
-				return_dict=False,
-			)
-
-		random_params = module_init_outputs["params"]
-
-		if params is not None:
-			random_params = flatten_dict(unfreeze(random_params))
-			params = flatten_dict(unfreeze(params))
-			for missing_key in self._missing_keys:
-				params[missing_key] = random_params[missing_key]
-			self._missing_keys = set()
-			return freeze(unflatten_dict(params))
-		else:
-			return random_params
-
-	def init_cache(self, batch_size, max_length):
-		"""The init_cache function is used to initialize the cache for a given batch size and sequence length.
-		The cache is a dictionary that contains all the intermediate states from each layer in the model.
-		This allows us to run inference on multiple batches without having to re-run forward passes through every layer in
-		the model, which would be very slow.
-
-		Args:
-		    self: Access the module
-		    batch_size: Define the batch size of the input tensors
-		    max_length: Set the length of the input sequence
-
-		Returns:
-		    A dictionary with the following keys:
-		"""
-
-		return super().init_cache(batch_size=batch_size, max_length=max_length)
-
-	def __call__(
-		self,
-		input_ids: Optional[chex.Array] = None,
-		input_embeds: Optional[chex.Array] = None,
-		attention_mask: Optional[chex.Array] = None,
-		position_ids: Optional[chex.Array] = None,
-		segment_ids: Optional[chex.Array] = None,
-		params: dict = None,
-		past_key_values: Optional[dict] = None,
-		dropout_rng: jax.random.PRNGKey = None,
-		train: bool = False,
-		output_attentions: Optional[bool] = None,
-		output_hidden_states: Optional[bool] = None,
-		return_dict: Optional[bool] = None,
-		add_params_field: bool = False,
-		**kwargs,
-	):
-		"""
-		Forward pass through the model.
-
-		Args:
-		    input_ids (chex.Array): Input tensor containing token IDs.
-		    input_embeds (Optional[chex.Array]): embedding inputs to be used instead of input_ids.
-		    attention_mask (Optional[chex.Array]): Mask for attention.
-		    position_ids (Optional[chex.Array]): Positional indices.
-		    segment_ids (Optional[chex.Array]): Segment IDs for distinguishing different parts of the input.
-		    params (dict, optional): Parameters for the model.
-		    past_key_values (dict, optional): Past key and value states for caching.
-		    dropout_rng (jax.random.PRNGKey, optional): RNG key for dropout.
-		    train (bool): If True, the model is in training mode.
-		    output_attentions (Optional[bool]): If True, output attention weights.
-		    output_hidden_states (Optional[bool]): If True, output hidden states.
-		    return_dict (Optional[bool]): If True, return a dictionary of outputs.
-		    add_params_field (bool): If True, include the parameters in the input dictionary.
-		    **kwargs: Additional arguments.
-
-		Returns:
-		    Output type depends on the model configuration.
-		"""
-		output_attentions = (
-			output_attentions
-			if output_attentions is not None
-			else self.config.output_attentions
-		)
-		output_hidden_states = (
-			output_hidden_states
-			if output_hidden_states is not None
-			else self.config.output_hidden_states
-		)
-		return_dict = return_dict if return_dict is not None else self.config.return_dict
-		batch_size, sequence_length = (
-			input_ids.shape if input_ids is not None else input_embeds.shape[:2]
-		)
-
-		if position_ids is None:
-			if past_key_values is not None:
-				raise ValueError(
-					"Make sure to provide `position_ids` when passing `past_key_values`."
-				)
-
-			position_ids = jnp.broadcast_to(
-				jnp.arange(sequence_length)[None, :], (batch_size, sequence_length)
-			)
-
-		if attention_mask is None:
-			attention_mask = jnp.ones((batch_size, sequence_length))
-
-		rngs = {}
-		if dropout_rng is not None:
-			rngs["dropout"] = dropout_rng
-
-		if self.config.bits is not None:
-			rngs["params"] = jax.random.key(0)
-
-		inputs = (
-			{"params": params or self.params} if add_params_field else params or self.params
-		)
-
-		if past_key_values is not None:
-			inputs["cache"] = past_key_values
-			mutable = ["cache"]
-		else:
-			mutable = False
-
-		outputs = self.module.apply(
-			inputs,
-			input_ids=jnp.array(input_ids, dtype="i4"),
-			attention_mask=jnp.array(attention_mask, dtype="i4"),
-			position_ids=jnp.array(position_ids, dtype="i4"),
-			deterministic=not train,
-			init_cache=False,
-			output_attentions=output_attentions,
-			output_hidden_states=output_hidden_states,
-			return_dict=return_dict,
-			input_embeds=input_embeds,
-			segment_ids=segment_ids,
-			rngs=rngs,
-			mutable=mutable,
-		)
-
-		if past_key_values is not None and return_dict:
-			outputs, past_key_values = outputs
-			outputs["past_key_values"] = unfreeze(past_key_values["cache"])
-			return outputs
-		elif past_key_values is not None and not return_dict:
-			outputs, past_key_values = outputs
-			outputs = outputs[:1] + (unfreeze(past_key_values["cache"]),) + outputs[1:]
-
-		return outputs
 
 
 class FlaxInternLM2BlockCollection(nn.Module):
@@ -743,7 +426,6 @@ class FlaxInternLM2BlockCollection(nn.Module):
 	def __call__(
 		self,
 		hidden_states: chex.Array,
-		frequencies: Tuple[chex.Array, chex.Array],
 		attention_mask: chex.Array,
 		causal_mask: chex.Array,
 		position_ids: chex.Array,
@@ -758,7 +440,6 @@ class FlaxInternLM2BlockCollection(nn.Module):
 
 		Args:
 		    hidden_states (chex.Array): Input tensor containing the hidden states.
-		    frequencies (Tuple[chex.Array, chex.Array]): Frequency positional encodings.
 		    attention_mask (chex.Array): Mask to apply during attention.
 		    causal_mask (chex.Array): Causal mask for autoregressive decoding.
 		    position_ids (chex.Array): Positional indices for the sequence.
@@ -804,7 +485,6 @@ class FlaxInternLM2BlockCollection(nn.Module):
 
 			layer_outputs = block(
 				hidden_states=hidden_states,
-				frequencies=frequencies,
 				attention_mask=attention_mask,
 				position_ids=position_ids,
 				causal_mask=causal_mask,
@@ -824,7 +504,14 @@ class FlaxInternLM2BlockCollection(nn.Module):
 		return outputs
 
 
-class FlaxInternLM2Module(nn.Module):
+@register_module(
+	"base-module",
+	config=InternLM2Config,
+	model_type="internlm2",
+	embedding_layer_names=["tok_embeddings"],
+)
+@wrap_easydel_module(config_class=InternLM2Config, base_model_prefix="model")
+class FlaxInternLM2Model(nn.Module):
 	config: InternLM2Config
 	dtype: jnp.dtype = jnp.float32
 	param_dtype: jnp.dtype = jnp.float32
@@ -850,48 +537,12 @@ class FlaxInternLM2Module(nn.Module):
 			dtype=self.dtype,
 			param_dtype=self.param_dtype,
 		)
-		config = self.config
 		self.causal_mask = nn.make_causal_mask(
 			jnp.ones(
 				shape=(1, self.config.granted_mask_max_position_embedding),
 				dtype="bool",
 			),
 			dtype="bool",
-		)
-
-		initial_rope_kwargs = dict(rope_type="none")
-		if config.rope_scaling is not None:
-			scaling_type = config.rope_scaling.get(
-				"type",
-				config.rope_scaling.get("rope_type"),
-			)
-			scaling_factor = config.rope_scaling.get("factor")
-
-			low_freq_factor = config.rope_scaling.get(
-				"low_freq_factor",
-				None,
-			)
-			high_freq_factor = config.rope_scaling.get(
-				"high_freq_factor",
-				None,
-			)
-			original_max_position_embeddings = config.rope_scaling.get(
-				"original_max_position_embeddings",
-				None,
-			)
-			initial_rope_kwargs = dict(
-				scaling_factor=scaling_factor,
-				rope_type=scaling_type,
-				low_freq_factor=low_freq_factor,
-				high_freq_factor=high_freq_factor,
-				original_max_position_embeddings=original_max_position_embeddings,
-			)
-
-		self.frequencies = precompute_frequencies(
-			max_position_embeddings=self.config.granted_freq_max_position_embedding,
-			dim=config.hidden_size // config.num_attention_heads,
-			base=config.rope_theta,
-			**initial_rope_kwargs,
 		)
 
 	def __call__(
@@ -941,7 +592,6 @@ class FlaxInternLM2Module(nn.Module):
 
 		outputs = self.layers(
 			hidden_states=input_embeds,
-			frequencies=self.frequencies,
 			attention_mask=attention_mask,
 			position_ids=position_ids,
 			causal_mask=self.causal_mask,
@@ -972,29 +622,20 @@ class FlaxInternLM2Module(nn.Module):
 
 
 @register_module(
-	"base-module",
+	"causal-language-model",
 	config=InternLM2Config,
 	model_type="internlm2",
 	embedding_layer_names=["tok_embeddings"],
 )
-class FlaxInternLM2Model(FlaxInternLM2PreTrainedModel):
-	module_class = FlaxInternLM2Module
-
-	def set_input_embeddings(self, value):
-		self.module.tok_embeddings = value
-
-	def get_input_embeddings(self):
-		return self.module.tok_embeddings
-
-
-class FlaxInternLM2ForCausalLMModule(nn.Module):
+@wrap_easydel_module(config_class=InternLM2Config, base_model_prefix="model")
+class FlaxInternLM2ForCausalLM(nn.Module):
 	config: InternLM2Config
 	dtype: jnp.dtype = jnp.float32
 	param_dtype: jnp.dtype = jnp.float32
 	precision: Optional[Union[jax.lax.Precision, str]] = None
 
 	def setup(self):
-		self.model = FlaxInternLM2Module(
+		self.model = FlaxInternLM2Model.flax_module(
 			self.config,
 			dtype=self.dtype,
 			param_dtype=self.param_dtype,
@@ -1090,75 +731,13 @@ class FlaxInternLM2ForCausalLMModule(nn.Module):
 
 
 @register_module(
-	"causal-language-model",
+	"sequence-classification",
 	config=InternLM2Config,
 	model_type="internlm2",
 	embedding_layer_names=["tok_embeddings"],
 )
-class FlaxInternLM2ForCausalLM(FlaxInternLM2PreTrainedModel):
-	module_class = FlaxInternLM2ForCausalLMModule
-
-	def set_input_embeddings(self, value):
-		self.module.model.tok_embeddings = value
-
-	def get_input_embeddings(self):
-		return self.module.model.tok_embeddings
-
-	def set_decoder(self, decoder):
-		self.module.model = decoder
-
-	def get_decoder(self):
-		return self.module.model
-
-	def get_output_embeddings(self):
-		return self.module.output
-
-	def set_output_embeddings(self, new_embeddings):
-		self.module.output = new_embeddings
-
-	def prepare_inputs_for_generation(
-		self, input_ids, max_length, attention_mask: Optional[chex.Array] = None
-	):
-		"""The prepare_inputs_for_generation function is used to prepare the inputs for a generation task.
-
-		Args:
-		    self: Access variables that belong to the class
-		    input_ids: Pass in the input tokens
-		    max_length: Set the length of the sequence to be generated
-		    attention_mask: Optional[chex.Array]: Mask the attention
-		        weights
-
-		Returns:
-		    A dictionary of the past_key_values, attention_mask and
-		    position ids
-		"""
-		batch_size, seq_length = input_ids.shape
-
-		past_key_values = self.init_cache(batch_size, max_length)
-		extended_attention_mask = jnp.ones((batch_size, max_length), dtype="i4")
-		if attention_mask is not None:
-			position_ids = attention_mask.cumsum(axis=-1) - 1
-			extended_attention_mask = lax.dynamic_update_slice(
-				extended_attention_mask, attention_mask, (0, 0)
-			)
-		else:
-			position_ids = jnp.broadcast_to(
-				jnp.arange(seq_length, dtype="i4")[None, :], (batch_size, seq_length)
-			)
-
-		return {
-			"past_key_values": past_key_values,
-			"attention_mask": extended_attention_mask,
-			"position_ids": position_ids,
-		}
-
-	def update_inputs_for_generation(self, model_outputs, model_kwargs):
-		model_kwargs["past_key_values"] = model_outputs.past_key_values
-		model_kwargs["position_ids"] = model_kwargs["position_ids"][:, -1:] + 1
-		return model_kwargs
-
-
-class FlaxInternLM2ForSequenceClassificationModule(nn.Module):
+@wrap_easydel_module(config_class=InternLM2Config, base_model_prefix="model")
+class FlaxInternLM2ForSequenceClassification(nn.Module):
 	num_classes: int
 	config: InternLM2Config
 	dtype: jnp.dtype = jnp.float32
@@ -1175,7 +754,12 @@ class FlaxInternLM2ForSequenceClassificationModule(nn.Module):
 		Returns:
 		    A tuple of the model and the classifier
 		"""
-		self.model = FlaxInternLM2Module(self.config, dtype=self.dtype)
+		self.model = FlaxInternLM2Model.flax_module(
+			self.config,
+			dtype=self.dtype,
+			param_dtype=self.param_dtype,
+			precision=self.precision,
+		)
 		self.score = Dense(
 			self.num_classes,
 			dtype=self.dtype,
@@ -1249,13 +833,3 @@ class FlaxInternLM2ForSequenceClassificationModule(nn.Module):
 			)
 		else:
 			return (prediction,)
-
-
-@register_module(
-	"sequence-classification",
-	config=InternLM2Config,
-	model_type="internlm2",
-	embedding_layer_names=["tok_embeddings"],
-)
-class FlaxInternLM2ForSequenceClassification(FlaxInternLM2PreTrainedModel):
-	module_class = FlaxInternLM2ForSequenceClassificationModule
