+++ easydel/modules/mistral/__init__.py
@@ -12,12 +12,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from easydel.modules.mistral.mistral_configuration import MistralConfig
+from easydel.modules.mistral.mistral_configuration import (
+	MistralConfig,
+	VisionMistralConfig,
+)
 from easydel.modules.mistral.modelling_mistral_flax import (
 	FlaxMistralForCausalLM,
-	FlaxMistralForCausalLMModule,
 	FlaxMistralModel,
-	FlaxMistralModule,
 	FlaxVisionMistralForCausalLM,
-	VisionMistralConfig,
+)
+
+__all__ = (
+	"FlaxMistralForCausalLM",
+	"FlaxMistralModel",
+	"FlaxVisionMistralForCausalLM",
+	"MistralConfig",
+	"VisionMistralConfig",
 )
