+++ easydel/modules/mistral/mistral_configuration.py
@@ -246,3 +246,42 @@ class MistralConfig(EasyDeLBaseConfig):
 			"mask_max_position_embeddings",
 			self.max_position_embeddings,
 		)
+
+
+class VisionMistralConfig(MistralConfig):
+	def __init__(
+		self,
+		vision_vocab_size=8448,
+		tie_vision_embeddings=False,
+		sample_mode="all",
+		**kwargs,
+	):
+		super().__init__(**kwargs)
+		self.vision_vocab_size = vision_vocab_size
+		self.tie_vision_embeddings = tie_vision_embeddings
+		self.sample_mode = sample_mode
+
+	def get_partition_rules(self, *args, **kwargs):
+		"""
+		Get the partition rules for the model.
+		Returns:
+		    `Tuple[Tuple[str, PartitionSpec]]`: The partition rules.
+		"""
+		return (
+			("model/embed_tokens/embedding", PartitionSpec("tp", ("fsdp", "sp"))),
+			("model/embed_vision/embedding", PartitionSpec("tp", ("fsdp", "sp"))),
+			(
+				"self_attn/(q_proj|k_proj|v_proj)/kernel",
+				PartitionSpec(("fsdp", "sp"), "tp"),
+			),
+			("self_attn/o_proj/kernel", PartitionSpec("tp", ("fsdp", "sp"))),
+			("mlp/gate_proj/kernel", PartitionSpec(("fsdp", "sp"), "tp")),
+			("mlp/down_proj/kernel", PartitionSpec("tp", ("fsdp", "sp"))),
+			("mlp/up_proj/kernel", PartitionSpec(("fsdp", "sp"), "tp")),
+			("input_layernorm/kernel", PartitionSpec(None)),
+			("post_attention_layernorm/kernel", PartitionSpec(None)),
+			("model/norm/kernel", PartitionSpec(None)),
+			("lm_head/kernel", PartitionSpec(("fsdp", "sp"), "tp")),
+			("vision_head/kernel", PartitionSpec(("fsdp", "sp"), "tp")),
+			(".*", PartitionSpec(None)),
+		)
