--- easydel/modules/mistral/vision_mistral_configuration.py
@@ -1,56 +1,0 @@
-# Copyright 2023 The EASYDEL Author @erfanzar (Erfan Zare Chavoshi).
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from jax.sharding import PartitionSpec
-
-from easydel.modules.mistral.mistral_configuration import MistralConfig as MistralConfig
-
-
-class VisionMistralConfig(MistralConfig):
-	def __init__(
-		self,
-		vision_vocab_size=8448,
-		tie_vision_embeddings=False,
-		sample_mode="all",
-		**kwargs,
-	):
-		super().__init__(**kwargs)
-		self.vision_vocab_size = vision_vocab_size
-		self.tie_vision_embeddings = tie_vision_embeddings
-		self.sample_mode = sample_mode
-
-	def get_partition_rules(self, *args, **kwargs):
-		"""
-		Get the partition rules for the model.
-		Returns:
-		    `Tuple[Tuple[str, PartitionSpec]]`: The partition rules.
-		"""
-		return (
-			("model/embed_tokens/embedding", PartitionSpec("tp", ("fsdp", "sp"))),
-			("model/embed_vision/embedding", PartitionSpec("tp", ("fsdp", "sp"))),
-			(
-				"self_attn/(q_proj|k_proj|v_proj)/kernel",
-				PartitionSpec(("fsdp", "sp"), "tp"),
-			),
-			("self_attn/o_proj/kernel", PartitionSpec("tp", ("fsdp", "sp"))),
-			("mlp/gate_proj/kernel", PartitionSpec(("fsdp", "sp"), "tp")),
-			("mlp/down_proj/kernel", PartitionSpec("tp", ("fsdp", "sp"))),
-			("mlp/up_proj/kernel", PartitionSpec(("fsdp", "sp"), "tp")),
-			("input_layernorm/kernel", PartitionSpec(None)),
-			("post_attention_layernorm/kernel", PartitionSpec(None)),
-			("model/norm/kernel", PartitionSpec(None)),
-			("lm_head/kernel", PartitionSpec(("fsdp", "sp"), "tp")),
-			("vision_head/kernel", PartitionSpec(("fsdp", "sp"), "tp")),
-			(".*", PartitionSpec(None)),
-		)
