+++ easydel/modules/modeling_utils.py
@@ -44,7 +44,7 @@ import jax.tree_util
 from fjformer.checkpoint import CheckpointManager
 from fjformer.dtypes import Array8Bit
 from fjformer.sharding import match_partition_rules
-from flax.core import FrozenDict, unfreeze
+from flax.core import FrozenDict, unfreeze, freeze
 from flax.traverse_util import flatten_dict, unflatten_dict
 from jax import numpy as jnp
 from jax.experimental.mesh_utils import create_device_mesh
@@ -64,6 +64,14 @@ from easydel.etils.etils import (
 )
 from easydel.etils.partition_module import PartitionAxis
 from easydel.utils.quantizers import DEFAULT_QUANTIZATION_PATTERN, EasyQuantizer
+
+import jax
+import jax.tree_util
+from jax import lax
+
+from easydel.generation.flax_utils import FlaxSampleOutput
+from easydel.generation.logits_process import FlaxLogitsProcessorList
+from easydel.inference.utils import SampleState
 
 logger = get_logger(__name__)
 
@@ -1870,6 +1878,493 @@ class EasyDeLBaseModule(FlaxPreTrainedModel):
 		return params
 
 
+class EasyDeLBaseVisionModule(EasyDeLBaseModule):
+	def init_cache(self, batch_size, max_length):
+		input_ids = jnp.ones((batch_size, max_length))
+		attention_mask = jnp.ones_like(input_ids)
+		position_ids = jnp.broadcast_to(
+			jnp.arange(jnp.atleast_2d(input_ids).shape[-1]), input_ids.shape
+		)
+		vision_mask = jnp.ones((batch_size, max_length), dtype=bool)
+
+		init_variables = self.module.init(
+			jax.random.PRNGKey(0),
+			input_ids=input_ids,
+			vision_mask=vision_mask,
+			attention_mask=attention_mask,
+			position_ids=position_ids,
+			return_dict=False,
+			init_cache=True,
+		)
+		return init_variables["cache"]
+
+	def init_weights(
+		self, rng: jax.random.PRNGKey, input_shape: Tuple, params: FrozenDict = None
+	) -> FrozenDict:
+		"""The init_weights function is used to initialize the weights of a model.
+
+		Args:
+		    self: Access variables that belong to the class
+		    rng: jax.random.PRNGKey: Initialize the weights of the model
+		    input_shape: Tuple: Specify the shape of the input tensor
+		    params: FrozenDict: Pass in the parameters of a pre-trained
+		        model
+
+		Returns:
+		    A frozendict of parameters
+		"""
+		input_ids = jnp.zeros(input_shape, dtype="i4")
+		attention_mask = jnp.ones_like(input_ids)
+		vision_mask = jnp.ones(input_ids.shape, dtype=bool)
+		position_ids = jnp.broadcast_to(
+			jnp.arange(jnp.atleast_2d(input_ids).shape[-1]),
+			input_shape,
+		)
+		params_rng, dropout_rng = jax.random.split(rng)
+
+		random_params = self.module.init(
+			{"params": params_rng, "dropout": dropout_rng},
+			input_ids=input_ids,
+			vision_mask=vision_mask,
+			attention_mask=attention_mask,
+			position_ids=position_ids,
+			return_dict=False,
+		)["params"]
+
+		if params is not None:
+			random_params = flatten_dict(unfreeze(random_params))
+			params = flatten_dict(unfreeze(params))
+			for missing_key in self._missing_keys:
+				params[missing_key] = random_params[missing_key]
+			self._missing_keys = set()
+			return freeze(unflatten_dict(params))
+		else:
+			return random_params
+
+	def __call__(
+		self,
+		input_ids: chex.Array,
+		vision_mask: Optional[chex.Array] = None,
+		attention_mask: Optional[chex.Array] = None,
+		position_ids: Optional[chex.Array] = None,
+		params: dict = None,
+		past_key_values: Optional[dict] = None,
+		dropout_rng: jax.random.PRNGKey = None,
+		train: bool = False,
+		output_attentions: Optional[bool] = None,
+		output_hidden_states: Optional[bool] = None,
+		return_dict: Optional[bool] = None,
+		extra_embedding: Optional[Union[jnp.ndarray, None]] = None,
+		add_params_field: bool = False,
+		**kwargs,
+	):
+		output_attentions = (
+			output_attentions
+			if output_attentions is not None
+			else self.config.output_attentions
+		)
+		output_hidden_states = (
+			output_hidden_states
+			if output_hidden_states is not None
+			else self.config.output_hidden_states
+		)
+		return_dict = return_dict if return_dict is not None else self.config.return_dict
+
+		batch_size, sequence_length = input_ids.shape
+
+		if position_ids is None:
+			if past_key_values is not None:
+				raise ValueError(
+					"Make sure to provide `position_ids` when passing `past_key_values`."
+				)
+
+			position_ids = jnp.broadcast_to(
+				jnp.arange(sequence_length)[None, :], (batch_size, sequence_length)
+			)
+
+		if attention_mask is None:
+			attention_mask = jnp.ones((batch_size, sequence_length))
+
+		rngs = {}
+		if dropout_rng is not None:
+			rngs["dropout"] = dropout_rng
+
+		inputs = (
+			{
+				"params": params or self.params,
+			}
+			if add_params_field
+			else params or self.params
+		)
+
+		if past_key_values is not None:
+			inputs["cache"] = past_key_values
+			mutable = ["cache"]
+		else:
+			mutable = False
+		kwargs.pop("deterministic", None)
+		kwargs.pop("init_cache", None)
+		child_call_args = dict(
+			input_ids=jnp.array(input_ids, dtype="i4"),
+			vision_mask=jnp.array(vision_mask, dtype="f4"),
+			attention_mask=jnp.array(attention_mask, dtype="i4"),
+			position_ids=jnp.array(position_ids, dtype="i4"),
+			deterministic=not train,
+			init_cache=False,
+			output_attentions=output_attentions,
+			output_hidden_states=output_hidden_states,
+			return_dict=return_dict,
+			extra_embedding=extra_embedding,
+			**kwargs,
+		)
+		all_kwargs = {k: v for k, v in child_call_args.items()}
+		filtered_kwargs = self._validate_signature(self.module.__call__, (), all_kwargs)
+		outputs = self.module.apply(inputs, rngs=rngs, mutable=mutable, **filtered_kwargs)
+
+		if past_key_values is not None and return_dict:
+			outputs, past_key_values = outputs
+			outputs["past_key_values"] = unfreeze(past_key_values["cache"])
+			return outputs
+		elif past_key_values is not None and not return_dict:
+			outputs, past_key_values = outputs
+			outputs = outputs[:1] + (unfreeze(past_key_values["cache"]),) + outputs[1:]
+
+		return outputs
+
+	def prepare_inputs_for_generation(
+		self,
+		input_ids: jax.Array,
+		max_length: int,
+		attention_mask: Optional[jax.Array] = None,
+		vision_mask: Optional[jax.Array] = None,
+	):
+		# initializing the cache
+		batch_size, seq_length = input_ids.shape
+
+		past_key_values = self.init_cache(batch_size, max_length)
+		extended_attention_mask = jnp.ones((batch_size, max_length), dtype="i4")
+		if attention_mask is not None:
+			position_ids = attention_mask.cumsum(axis=-1) - 1
+			extended_attention_mask = lax.dynamic_update_slice(
+				extended_attention_mask, attention_mask, (0, 0)
+			)
+		else:
+			position_ids = jnp.broadcast_to(
+				jnp.arange(seq_length, dtype="i4")[None, :], (batch_size, seq_length)
+			)
+
+		return {
+			"past_key_values": past_key_values,
+			"attention_mask": extended_attention_mask,
+			"position_ids": position_ids,
+			"vision_mask": vision_mask,
+		}
+
+	def update_inputs_for_generation(self, model_outputs, model_kwargs):
+		return {
+			"past_key_values": model_outputs.past_key_values,
+			"position_ids": model_kwargs["position_ids"][:, -1:] + 1,
+			"attention_mask": model_kwargs["attention_mask"],
+			"vision_mask": model_kwargs["vision_mask"],
+		}
+
+	def _sample_vision(
+		self,
+		input_ids: None,
+		max_length: Optional[int] = None,
+		pad_token_id: Optional[int] = None,
+		eos_token_id: Optional[int] = None,
+		prng_key: Optional[jnp.ndarray] = None,
+		logits_processor: Optional[FlaxLogitsProcessorList] = None,
+		logits_warper: Optional[FlaxLogitsProcessorList] = None,
+		cfg_scales: jnp.ndarray = 1.0,
+		trace: bool = True,
+		params: Optional[Dict[str, jnp.ndarray]] = None,
+		model_kwargs: Optional[Dict[str, jnp.ndarray]] = None,
+	):
+		# init values
+		max_length = (
+			max_length if max_length is not None else self.generation_config.max_length
+		)
+		pad_token_id = (
+			pad_token_id if pad_token_id is not None else self.generation_config.pad_token_id
+		)
+		eos_token_id = (
+			eos_token_id if eos_token_id is not None else self.generation_config.eos_token_id
+		)
+		prng_key = prng_key if prng_key is not None else jax.random.PRNGKey(0)
+
+		batch_size, cur_len = input_ids.shape
+		initial_len = cur_len
+
+		eos_token_id = jnp.array(
+			eos_token_id, dtype=jnp.int32 if eos_token_id is not None else None
+		)
+		pad_token_id = jnp.array(pad_token_id, dtype=jnp.int32)
+		cur_len = jnp.array(cur_len)
+
+		# per batch-item holding current token in loop.
+		sequences = jnp.full((batch_size, max_length), pad_token_id, dtype=jnp.int32)
+		sequences = lax.dynamic_update_slice(sequences, input_ids, (0, 0))
+
+		# per batch-item state bit indicating if sentence has finished.
+		is_sent_finished = jnp.zeros((batch_size,), dtype=jnp.bool_)
+
+		# For Seq2Seq generation, we only need to use the decoder instead of the whole model in generation loop
+		# and pass it the `encoder_outputs`, which are part of the `model_kwargs`.
+		model = self.decode if self.config.is_encoder_decoder else self
+
+		# initialize model specific kwargs
+		model_kwargs = self.prepare_inputs_for_generation(
+			input_ids, max_length, **model_kwargs
+		)
+
+		# initialize state
+		state = SampleState(
+			cur_len=cur_len,
+			sequences=sequences,
+			running_token=input_ids,
+			is_sent_finished=is_sent_finished,
+			prng_key=prng_key,
+			model_kwargs=model_kwargs,
+		)
+
+		def sample_search_cond_fn(state):
+			"""state termination condition fn."""
+			has_reached_max_length = state.cur_len == max_length
+			all_sequence_finished = jnp.all(state.is_sent_finished)
+			finish_generation = jnp.logical_or(has_reached_max_length, all_sequence_finished)
+			return ~finish_generation
+
+		def sample_search_body_fn(state):
+			"""state update fn."""
+			prng_key, prng_key_next = jax.random.split(state.prng_key)
+			model_outputs = model(state.running_token, params=params, **state.model_kwargs)
+
+			logits = model_outputs.logits[:, -1]
+			cond_logits, uncond_logits = jnp.split(logits, 2, axis=0)
+			logits = uncond_logits + cfg_scales[:, None] * (cond_logits - uncond_logits)
+
+			# apply min_length, ...
+			logits = logits_processor(state.sequences, logits, state.cur_len)
+			# apply top_p, top_k, temperature
+			logits = logits_warper(logits, logits, state.cur_len)
+
+			next_token = jax.random.categorical(prng_key, logits, axis=-1)
+			next_token = jax.lax.cond(
+				(state.cur_len - initial_len + 1) % 257 == 0,
+				lambda: jnp.full_like(next_token, 8192),
+				lambda: next_token,
+			)
+			next_token = jnp.concatenate([next_token, next_token], axis=0)
+
+			# next_token = next_token * ~state.is_sent_finished + pad_token_id * state.is_sent_finished
+			next_is_sent_finished = state.is_sent_finished | (next_token == eos_token_id)
+			next_token = next_token[:, None]
+
+			next_sequences = lax.dynamic_update_slice(
+				state.sequences, next_token, (0, state.cur_len)
+			)
+			next_model_kwargs = self.update_inputs_for_generation(
+				model_outputs, state.model_kwargs
+			)
+
+			return SampleState(
+				cur_len=state.cur_len + 1,
+				sequences=next_sequences,
+				running_token=next_token,
+				is_sent_finished=next_is_sent_finished,
+				model_kwargs=next_model_kwargs,
+				prng_key=prng_key_next,
+			)
+
+		if input_ids.shape[1] > 1:
+			state = sample_search_body_fn(state)
+
+		if not trace:
+			state = self._run_loop_in_debug(
+				sample_search_cond_fn, sample_search_body_fn, state
+			)
+		else:
+			state = lax.while_loop(sample_search_cond_fn, sample_search_body_fn, state)
+
+		return FlaxSampleOutput(sequences=state.sequences)
+
+	def generate_vision(
+		self,
+		input_ids: jnp.ndarray,
+		cfg_scales: jnp.ndarray,
+		generation_config: Optional["transformers.GenerationConfig"] = None,  # noqa #type:ignore
+		prng_key: Optional[jnp.ndarray] = None,
+		trace: bool = True,
+		params: Optional[Dict[str, jnp.ndarray]] = None,
+		logits_processor: Optional[FlaxLogitsProcessorList] = None,
+		**kwargs,
+	):
+		self._validate_model_class()
+
+		if generation_config is None:
+			if (
+				self.generation_config._from_model_config
+				and self.generation_config._original_object_hash == hash(self.generation_config)
+			):
+				from transformers import GenerationConfig
+
+				new_generation_config = GenerationConfig.from_model_config(self.config)
+				if new_generation_config != self.generation_config:
+					logger.warn(
+						"You have modified the pretrained model configuration to control generation. This is a"
+						" deprecated strategy to control generation and will be removed soon, in a future version."
+						" Please use and modify the model generation configuration (see"
+						" https://huggingface.co/docs/transformers/generation_strategies#"
+						"default-text-generation-configuration )"
+					)
+					self.generation_config = new_generation_config
+			generation_config = self.generation_config
+		import copy
+
+		generation_config = copy.deepcopy(generation_config)
+		model_kwargs = generation_config.update(
+			**kwargs
+		)  # All unused kwargs must be model kwargs
+		generation_config.validate()
+		self._validate_model_kwargs(model_kwargs.copy())
+
+		logits_processor = (
+			logits_processor if logits_processor is not None else FlaxLogitsProcessorList()
+		)
+
+		# set init values
+		prng_key = prng_key if prng_key is not None else jax.random.PRNGKey(0)
+
+		if (
+			generation_config.pad_token_id is None
+			and generation_config.eos_token_id is not None
+		):
+			if model_kwargs.get("attention_mask") is None:
+				logger.warn(
+					"The attention mask and the pad token id were not set. As a consequence, you may observe "
+					"unexpected behavior. Please pass your input's `attention_mask` to obtain reliable results."
+				)
+			eos_token_id = generation_config.eos_token_id
+			if isinstance(eos_token_id, list):
+				eos_token_id = eos_token_id[0]
+			logger.warn(
+				f"Setting `pad_token_id` to `eos_token_id`:{eos_token_id} for open-end generation."
+			)
+			generation_config.pad_token_id = eos_token_id
+
+		if (
+			generation_config.decoder_start_token_id is None
+			and self.config.is_encoder_decoder
+		):
+			raise ValueError(
+				"`decoder_start_token_id` has to be defined for encoder-decoder generation."
+			)
+
+		# decoder-only models should use left-padding for generation (can't be checked with `trace=True`)
+		if not self.config.is_encoder_decoder and not trace:
+			if (
+				generation_config.pad_token_id is not None
+				and jnp.sum(input_ids[:, -1] == generation_config.pad_token_id) > 0
+			):
+				logger.warn(
+					"A decoder-only architecture is being used, but right-padding was detected! For correct "
+					"generation results, please set `padding_side='left'` when initializing the tokenizer."
+				)
+
+		batch_size = input_ids.shape[0]
+
+		if self.config.is_encoder_decoder:
+			# add encoder_outputs to model_kwargs
+			if model_kwargs.get("encoder_outputs") is None:
+				model_kwargs = self._prepare_encoder_decoder_kwargs_for_generation(
+					input_ids, params, model_kwargs
+				)
+			# prepare decoder_input_ids for generation
+			input_ids = self._prepare_decoder_input_ids_for_generation(
+				batch_size,
+				decoder_start_token_id=generation_config.decoder_start_token_id,
+				bos_token_id=generation_config.bos_token_id,
+				model_kwargs=model_kwargs,
+			)
+
+		# Prepare `max_length` depending on other stopping criteria.
+		input_ids_seq_length = input_ids.shape[-1]
+		has_default_max_length = (
+			kwargs.get("max_length") is None and generation_config.max_length is not None
+		)
+		if (
+			has_default_max_length
+			and generation_config.max_new_tokens is None
+			and generation_config.max_length == 20
+		):
+			# 20 is the default max_length of the generation config
+			logger.warn(
+				f"Using the model-agnostic default `max_length` (={generation_config.max_length}) "
+				"to control the generation length.  recommend setting `max_new_tokens` to control"
+				" the maximum length of the generation.",
+				UserWarning,
+			)
+		elif generation_config.max_new_tokens is not None:
+			if not has_default_max_length and generation_config.max_length is not None:
+				logger.warn(
+					f"Both `max_new_tokens` (={generation_config.max_new_tokens}) and `max_length`(="
+					f"{generation_config.max_length}) seem to have been set. `max_new_tokens` will take precedence. "
+					"Please refer to the documentation for more information. "
+					"(https://huggingface.co/docs/transformers/main/en/main_classes/text_generation)"
+				)
+			generation_config.max_length = (
+				generation_config.max_new_tokens + input_ids_seq_length
+			)
+
+		if (
+			generation_config.min_length is not None
+			and generation_config.min_length > generation_config.max_length
+		):
+			raise ValueError(
+				f"Unfeasable length constraints: the minimum length ({generation_config.min_length}) is larger than"
+				f" the maximum length ({generation_config.max_length})"
+			)
+		if input_ids_seq_length >= generation_config.max_length:
+			input_ids_string = (
+				"decoder_input_ids" if self.config.is_encoder_decoder else "input_ids"
+			)
+			logger.warn(
+				f"Input length of {input_ids_string} is {input_ids_seq_length}, but `max_length` is set to"
+				f" {generation_config.max_length}. This can lead to unexpected behavior. You should consider"
+				" increasing`max_new_tokens`."
+			)
+
+		logits_processor = self._get_logits_processor(
+			generation_config=generation_config,
+			input_ids_seq_length=input_ids_seq_length,
+			logits_processor=logits_processor,
+		)
+
+		if not generation_config.do_sample and generation_config.num_beams == 1:
+			raise NotImplementedError
+		elif generation_config.do_sample and generation_config.num_beams == 1:
+			logits_warper = self._get_logits_warper(generation_config=generation_config)
+			return self._sample_vision(
+				input_ids,
+				generation_config.max_length,
+				generation_config.pad_token_id,
+				generation_config.eos_token_id,
+				prng_key,
+				logits_warper=logits_warper,
+				logits_processor=logits_processor,
+				cfg_scales=cfg_scales,
+				trace=trace,
+				params=params,
+				model_kwargs=model_kwargs,
+			)
+		elif not generation_config.do_sample and generation_config.num_beams > 1:
+			raise NotImplementedError
+		else:
+			raise NotImplementedError("`Beam sampling is currently not implemented.")
+
+
 def wrap_easydel_module(
 	config_class: Type[EasyDeLBaseConfig],
 	base_model_prefix: str = "model",
@@ -1893,6 +2388,39 @@ def wrap_easydel_module(
 				class_dict[name] = attr
 
 		WrappedModule = type(mdl.__name__, (EasyDeLBaseModule,), class_dict)
+		WrappedModule.__module__ = mdl.__module__
+		WrappedModule.__qualname__ = mdl.__qualname__
+		WrappedModule.__doc__ = mdl.__doc__
+
+		return WrappedModule
+
+	return wrapper
+
+
+def wrap_custom_easydel_module(
+	base,
+	config_class: Type[EasyDeLBaseConfig],
+	base_model_prefix: str = "model",
+):
+	def wrapper(mdl: Type[M]) -> Type[EasyDeLBaseModule]:
+		class_dict = {
+			"config_class": config_class,
+			"base_model_prefix": base_model_prefix,
+			"module_class": mdl,
+			"flax_module": mdl,
+			"__annotations__": {
+				"config_class": Type[EasyDeLBaseConfig],
+				"base_model_prefix": str,
+				"flax_module": Type[M],
+				"module_class": Union[flax.linen.Module, Type[M]],
+			},
+		}
+
+		for name, attr in mdl.__dict__.items():
+			if not name.startswith("__"):
+				class_dict[name] = attr
+
+		WrappedModule = type(mdl.__name__, (base,), class_dict)
 		WrappedModule.__module__ = mdl.__module__
 		WrappedModule.__qualname__ = mdl.__qualname__
 		WrappedModule.__doc__ = mdl.__doc__
