+++ easydel/modules/mosaic_mpt/modelling_mpt_flax.py
@@ -221,9 +221,9 @@ class FlaxMptAttention(FlaxAttentionModule):
 
 		if self.has_variable("cache", "cached_key") or init_cache:
 			key_states, value_states, attention_mask = self._concatenate_to_cache(
+				query_states,
 				key_states,
 				value_states,
-				query_states,
 				attention_mask,
 			)
 		if position_bias is not None:
