+++ easydel/modules/olmo/modelling_olmo_flax.py
@@ -329,9 +329,9 @@ class FlaxOlmoAttention(FlaxAttentionModule):
 			dropout_rng = self.make_rng("dropout")
 		if self.has_variable("cache", "cached_key") or init_cache:
 			key_states, value_states, attention_mask = self._concatenate_to_cache(
+				query_states,
 				key_states,
 				value_states,
-				query_states,
 				attention_mask,
 			)
