+++ easydel/modules/opt/modelling_opt_flax.py
@@ -184,9 +184,9 @@ class FlaxOPTAttention(FlaxAttentionModule):
 
 		if self.causal and (self.has_variable("cache", "cached_key") or init_cache):
 			key_states, value_states, attention_mask = self._concatenate_to_cache(
+				query_states,
 				key_states,
 				value_states,
-				query_states,
 				attention_mask,
 			)
 			if attention_mask is not None:
