+++ easydel/modules/phi/modelling_phi_flax.py
@@ -351,9 +351,9 @@ class FlaxPhiAttention(FlaxAttentionModule):
 
 		if self.has_variable("cache", "cached_key") or init_cache:
 			key_states, value_states, attention_mask = self._concatenate_to_cache(
+				query_states,
 				key_states,
 				value_states,
-				query_states,
 				attention_mask,
 			)
