+++ easydel/modules/phi3/modelling_phi3_flax.py
@@ -322,9 +322,9 @@ class FlaxPhi3Attention(FlaxAttentionModule):
 
 		if self.has_variable("cache", "cached_key") or init_cache:
 			key_states, value_states, attention_mask = self._concatenate_to_cache(
+				query_states,
 				key_states,
 				value_states,
-				query_states,
 				attention_mask,
 			)
