+++ easydel/modules/phimoe/modeling_phimoe_flax.py
@@ -314,9 +314,9 @@ class FlaxPhiMoEAttention(FlaxAttentionModule):
 
 		if self.has_variable("cache", "cached_key") or init_cache:
 			key_states, value_states, attention_mask = self._concatenate_to_cache(
+				query_states,
 				key_states,
 				value_states,
-				query_states,
 				attention_mask,
 			)
