+++ easydel/modules/qwen1/modeling_qwen1_flax.py
@@ -380,9 +380,9 @@ class FlaxQwen1Attention(FlaxAttentionModule):
 
 		if self.has_variable("cache", "cached_key") or init_cache:
 			key_states, value_states, attention_mask = self._concatenate_to_cache(
+				query_states,
 				key_states,
 				value_states,
-				query_states,
 				attention_mask,
 			)
