+++ easydel/modules/qwen2_moe/modeling_qwen2_moe_flax.py
@@ -387,9 +387,9 @@ class FlaxQwen2MoeAttention(FlaxAttentionModule):
 
 		if self.has_variable("cache", "cached_key") or init_cache:
 			key_states, value_states, attention_mask = self._concatenate_to_cache(
+				query_states,
 				key_states,
 				value_states,
-				query_states,
 				attention_mask,
 			)
