+++ easydel/modules/roberta/modelling_roberta_flax.py
@@ -240,9 +240,9 @@ class FlaxRobertaSelfAttention(FlaxAttentionModule):
 
 		if self.causal and (self.has_variable("cache", "cached_key") or init_cache):
 			key_states, value_states, attention_mask = self._concatenate_to_cache(
+				query_states,
 				key_states,
 				value_states,
-				query_states,
 				attention_mask,
 			)
