+++ easydel/modules/stablelm/modelling_stablelm_flax.py
@@ -439,9 +439,9 @@ class FlaxStableLmAttention(FlaxAttentionModule):
 			dropout_rng = self.make_rng("dropout")
 		if self.has_variable("cache", "cached_key") or init_cache:
 			key_states, value_states, attention_mask = self._concatenate_to_cache(
+				query_states,
 				key_states,
 				value_states,
-				query_states,
 				attention_mask,
 			)
