+++ easydel/modules/whisper/modelling_whisper_flax.py
@@ -177,9 +177,9 @@ class FlaxWhisperAttention(FlaxAttentionModule):
 
 		if self.causal and (self.has_variable("cache", "cached_key") or init_cache):
 			key_states, value_states, attention_mask = self._concatenate_to_cache(
+				query_states,
 				key_states,
 				value_states,
-				query_states,
 				attention_mask,
 			)
