+++ easydel/modules/xerxes/modeling_xerxes_flax.py
@@ -272,9 +272,9 @@ class FlaxXerxesAttention(FlaxAttentionModule):
 
 		if self.has_variable("cache", "cached_key") or init_cache:
 			key_states, value_states, attention_mask = self._concatenate_to_cache(
+				query_states,
 				key_states,
 				value_states,
-				query_states,
 				attention_mask,
 			)
